Fix up EC2 case and extract shared code and test
